change the execute order. now a savefile name should be selected by static method in QFileDialog, then the log file will be generated.
also, the generation of the log file is changed, now it runs the loop over generator->write sequence.

before, it has to generate all log in memory, then write to a file.

another important thing is the angle for specified point now is generated by preceding point and the following point,
but before it is generate by the current point and the following point. This at least improves the accuracy for a short route

but there must be a coincidence that the preceding point and the following point is the same one, due to all points slicing in map.py.
this problem may occur no matter how we calculate the angle.

the solution could be if the preceding point is the same as the following point, then we use the current point and the following point to calcuate the angle instead. this solution should be implemented in next update.